Fix the mobileState store behavior:
Export a $isMobile atom by convention (the $ for "store")
as we do in themeState.ts to be consistent.  Also, since
the media query string in TailwindCSS is
"(min-width: 768px)" instead of "(max-width: 768px)", the logic
is reversed, and $isMobile should be true when it DOESN'T
match, so the logic is updated now.

Update the componets to listen to $isMobile and re-render
accordingly.