Add a form handler for sending Nostr messages:
Use the nostr-dev-kit library to create
generate an ephemeral  Nostr keypair used to
sign and send to a set of relays.  Use my Nostr
npub to encrypt the message using the NIP-44
scheme (fixes NIP-04 encryption deficiencies).

Use a singleton pattern so multiple messages can be
sent with the same session keypair, despite it being
ephemeral.

Look Ma', no SMTP. Thank you Nostr.

The test is minimal since the NDKEvent is encrypted,
so testing the contents isn't possible without the
nsec of my npub.  At least test that the event is created correctly.